+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (C) 2005-2013 Leo Feyer
+ *
+ * @package   fen
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2013
+ */
+
+/**
+ * legends
+ */
+$GLOBALS['TL_LANG']['tl_settings']['teamtournament_legend']                      = 'Mannschaftsturniere';
+
+/**
+ * fields
+ */
+$GLOBALS['TL_LANG']['tl_settings']['teamtournament_defaultImageMen'] = array('Standardbild Männer', 'Standardbild für Listen mit männlichen Spielern');
+$GLOBALS['TL_LANG']['tl_settings']['teamtournament_defaultImageWomen'] = array('Standardbild Frauen', 'Standardbild für Listen mit weiblichen Spielern.');