+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Mannschaftsturniere-Bundle Changelog
 
+## Version 0.1.3 (2024-02-12)
+
+* Add: CSS-Klassen bei Paarungslisten erweitert
+* Add: Leerzeile zwischen Paarungen
+* Change: CSS-Klasse elo -> rating
+* Change: Paarungsliste nach Tischreihenfolge statt alphabetisch nach Mannschaft 1
+* Add: tl_settings mit Einstellungen Standardbild Frauen und Männer
+* Add: tl_teamtournament.gender -> für Abfrage ob Männer- oder Frauenturnier
+* Add: tl_teamtournament_teams.flag -> Dateiauswahl für ein Bild der Mannschaft, z.B. ein Flaggen-Symbol
+* Add: Bilder für Mannschaften
+* Add: Bildgröße tl_teamtournament.imageSize_flags für Flaggen der Mannschaften
+
 ## Version 0.1.2 (2024-02-11)
 
 * Add: Tabellen für Wettkämpfe und die Bretter
- 
+
 ## Version 0.1.1 (2024-02-03)
 
 * Fix: Tabelle ohne Kopfzeile